move sorted_by_key to NonEmptyIterator
